fix: cannot pulse on a non-reference type

+++ Assets/Scripts/Backend/Tests/Engine/EngineTest.cs
@@ -16,7 +16,7 @@ public class EngineTest
         Thread.Sleep(sleepTimeMs);
 
         var expectedTicks = Convert.ToInt32(
-            (double)ticksPerSecond * ((double)sleepTimeMs / 1000.0)
+            ticksPerSecond * (sleepTimeMs / 1000.0)
         );
         
         var timeAfter = engine.SimulationTime;
